# swagger.yaml
openapi: 3.0.0
info:
  title: Vet API
  contact: {}
  version: '1.0'
servers:
- url: https://vet-api-jik.herokuapp.com/api/v1
  variables: {}
paths:
...
  /auth/{id}:
    put:
      tags:
      - Auth
      summary: Update User
      operationId: UpdateUser
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: 619c01c44381cac63de3733a
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            example:
              username: KevinStudent
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - Auth
      summary: Delete User
      operationId: DeleteUser
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: 619c15a21300d5bbcd5a0cef
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    UpdateUserRequest:
      title: UpdateUserRequest
      required:
      - username
      type: object
      properties:
        username:
          type: string
      example:
        username: KevinStudent
...
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...